requirements.txt
functions-framework==3.*
gspread==5.4.0
pandas==1.4.2
numpy==1.22.0

google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-api-python-client==2.101.0